use the mount NS instead of the PID NS
To discriminate processes inside the container from those outside of it,
use the inum of the mount namespace of the container's init process
instead of the PID namespace.

Using the PID namespace is causing issues when running Podman in MPI
where containers share the PID NS with the host to allow for a proper
IPC among the MPI instances.

Fixes: #36
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>